requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-core
fastapi
python-dotenv
uvicorn
ddgs